requirements.txt
PyYAML
watchdog
rich
